# openapi.yaml
openapi: 3.0.3
info:
  title: Sunset Beach Resort & Spa — Admin/Booking API
  description: >
    Contract reverse-engineered from the current Next.js route handlers under
    `app/api/**/route.ts`. It documents exactly what each route accepts and
    returns today (including quirks such as Prisma `Decimal`/`Date` JSON
    serialization), so it can serve as a migration contract for a backend
    rewrite in another language. No behavior was changed to produce this spec.
  version: "1.0.0"
  license:
    name: Proprietary
  x-codegen-notes: >
    A property marked `x-sensitive: true` (guest email/phone, payment notes) is redacted as
    "[REDACTED]" in its generated model's toString() by the customized pojo.mustache template
    under openapi-templates/ - see scripts/generate-api.sh, which passes --template-dir. This is
    the one and only place PII redaction needs to be declared; it survives regeneration
    automatically. Do not hand-edit a generated model's toString() to redact a field instead -
    mark the field here and regenerate.
servers:
  - url: /api
    description: Relative to the deployed Next.js app origin.
...
security:
  - bearerAuth: []
paths:
...
  /public/rooms/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublicRoom
      tags: [Public]
      summary: Get a room by id (public)
      description: Unauthenticated. Powers the public room detail page.
      security: []
      responses:
        "200":
          description: The room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
  /public/rooms/{id}/pricing:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublicRoomPricing
      tags: [Public]
      summary: Get per-day prices for a room for a given month (public)
      description: Unauthenticated. Same shape as the staff endpoint - `isOverride` isn't sensitive, it just means the day has an explicit rate.
      security: []
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: 'Month key `YYYY-MM`. Defaults to the current UTC month.'
      responses:
        "200":
          description: Base price and the resolved price for every day of the month.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingResponse"
        "404":
          description: Room not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
  /public/rooms/{id}/availability:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublicRoomAvailability
      tags: [Public]
      summary: Get day-by-day availability for a room for a given month (public)
      description: >
        Unauthenticated. Powers the guest booking calendar. Unlike the staff
        endpoint, this never reveals *why* a day is blocked (real booking vs.
        manual staff block), nor how many units remain (`availableCount`) -
        only whether at least one unit is still bookable. Exposing the exact
        remaining count would let competitors read occupancy/load straight
        off the public site.
      security: []
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: 'Month key `YYYY-MM`. Defaults to the current UTC month.'
      responses:
        "200":
          description: Availability for every day of the month.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAvailabilityResponse"
        "404":
          description: Room not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Room not found
  /uploads/rooms/{roomId}/{filename}:
    parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
      - name: filename
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getRoomImage
      tags: [Public]
      summary: Serve an uploaded room image
      description: >
        Unauthenticated. Serves files written under `UPLOADS_ROOT/rooms/{roomId}/`
        by `POST /rooms/{id}/images`. The path shape matches `Room.images`
        entries exactly (`/uploads/rooms/{id}/{filename}`); on the deployed
        VPS the reverse proxy in front of this service rewrites public
        `/uploads/...` requests to this app's `/api/uploads/...` route so
        `Room.images` values need no prefix rewriting on the frontend.
      security: []
      responses:
        "200":
          description: The image file.
          content:
            image/jpeg: { schema: { type: string, format: binary } }
            image/png: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        "404":
          description: No such file (unknown room, unknown filename, or a path escaping the room's upload directory).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
              example:
                error: Not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT returned by `POST /auth/login`, sent as `Authorization: Bearer
        <token>`. Contains `sub` (user id), `email`, `role`, standard `iat`/`exp`
        (default TTL 7 days, configurable via `JWT_TTL_DAYS`). Every route in
        this file except `POST /bookings` and the unauthenticated `/auth/*`
        endpoints (`login`, `logout`) requires a valid token, returning 401 if
        missing/invalid/expired, and 403 where the route additionally requires
        role `ADMIN`.
...
  schemas:
...
    ErrorMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
      description: 'Generic error shape used by all hand-written `{ error: "..." }` responses.'
...
    Room:
      type: object
      description: >
        Despite the name this is a **room type**, not a single physical room — individual
        sellable units are separate `RoomUnit` rows (see the RoomUnits tag). `activeUnitCount`
        is computed from those rows (`count(RoomUnit WHERE roomId = id AND isActive)`), not
        stored. `basePrice` is a Prisma `Decimal` — it serializes to a **string**, not a JSON
        number. `createdAt` serializes as an ISO-8601 datetime string.
      required: [id, name, description, capacity, activeUnitCount, basePrice, images, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        capacity:
          type: integer
        activeUnitCount:
          type: integer
          minimum: 0
          description: 'Computed - number of active (`isActive: true`) `RoomUnit` rows of this type.'
        basePrice:
          type: string
          description: Decimal(10,2) rendered as a string, e.g. `"1500.00"`.
        images:
          type: array
          items:
            type: string
          description: Paths under `/uploads/rooms/{id}/...`.
        createdAt:
          type: string
          format: date-time
...
    PricingDay:
      type: object
      required: [date, price, isOverride]
      properties:
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        price:
          type: number
          description: Explicitly converted with `Number(...)` server-side — a real JSON number, unlike `Room.basePrice`.
        isOverride:
          type: boolean
          description: 'True if this day has an explicit `RatePlan` row; false if it fell back to `Room.basePrice`.'
    PricingResponse:
      type: object
      required: [basePrice, days]
      properties:
        basePrice:
          type: number
          description: 'Room.basePrice explicitly converted with `Number(...)`.'
        days:
          type: array
          items:
            $ref: "#/components/schemas/PricingDay"
...
    PublicAvailabilityDay:
      type: object
      description: >
        Like `AvailabilityDay` but collapsed to a yes/no signal - guests get
        neither the block/booking split nor the exact remaining count
        (revealing occupancy load to competitors), just whether at least one
        unit is still sellable.
      required: [date, isBlocked]
      properties:
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        isBlocked:
          type: boolean
          description: 'True when availableCount <= 0 for this day.'
    PublicAvailabilityResponse:
      type: object
      required: [days]
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/PublicAvailabilityDay"
...