# src/main/java/com/sunsetbeach/api/PublicApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sunsetbeach.api;

import com.sunsetbeach.model.ErrorMessage;
import com.sunsetbeach.model.PricingResponse;
import com.sunsetbeach.model.PublicAvailabilityResponse;
import com.sunsetbeach.model.Room;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
public interface PublicApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /public/rooms/{id} : Get a room by id (public)
     * Unauthenticated. Powers the public room detail page.
     *
     * @param id  (required)
     * @return The room. (status code 200)
     *         or Room not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/public/rooms/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Room> getPublicRoom(
         @PathVariable("id") String id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /public/rooms/{id}/availability : Get day-by-day availability for a room for a given month (public)
     * Unauthenticated. Powers the guest booking calendar. Unlike the staff endpoint, this never reveals *why* a day is blocked (real booking vs. manual staff block), nor how many units remain (&#x60;availableCount&#x60;) - only whether at least one unit is still bookable. Exposing the exact remaining count would let competitors read occupancy/load straight off the public site. 
     *
     * @param id  (required)
     * @param month Month key &#x60;YYYY-MM&#x60;. Defaults to the current UTC month. (optional)
     * @return Availability for every day of the month. (status code 200)
     *         or Room not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/public/rooms/{id}/availability",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PublicAvailabilityResponse> getPublicRoomAvailability(
         @PathVariable("id") String id,
        @Pattern(regexp = "^\\d{4}-\\d{2}$")  @Valid @RequestParam(value = "month", required = false) String month
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"days\" : [ { \"date\" : \"date\", \"isBlocked\" : true }, { \"date\" : \"date\", \"isBlocked\" : true } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /public/rooms/{id}/pricing : Get per-day prices for a room for a given month (public)
     * Unauthenticated. Same shape as the staff endpoint - &#x60;isOverride&#x60; isn&#39;t sensitive, it just means the day has an explicit rate.
     *
     * @param id  (required)
     * @param month Month key &#x60;YYYY-MM&#x60;. Defaults to the current UTC month. (optional)
     * @return Base price and the resolved price for every day of the month. (status code 200)
     *         or Room not found. (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/public/rooms/{id}/pricing",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PricingResponse> getPublicRoomPricing(
         @PathVariable("id") String id,
        @Pattern(regexp = "^\\d{4}-\\d{2}$")  @Valid @RequestParam(value = "month", required = false) String month
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"days\" : [ { \"date\" : \"date\", \"price\" : 6.027456183070403, \"isOverride\" : true }, { \"date\" : \"date\", \"price\" : 6.027456183070403, \"isOverride\" : true } ], \"basePrice\" : 0.8008281904610115 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /uploads/rooms/{roomId}/{filename} : Serve an uploaded room image
     * Unauthenticated. Serves files written under &#x60;UPLOADS_ROOT/rooms/{roomId}/&#x60; by &#x60;POST /rooms/{id}/images&#x60;. The path shape matches &#x60;Room.images&#x60; entries exactly (&#x60;/uploads/rooms/{id}/{filename}&#x60;); on the deployed VPS the reverse proxy in front of this service rewrites public &#x60;/uploads/...&#x60; requests to this app&#39;s &#x60;/api/uploads/...&#x60; route so &#x60;Room.images&#x60; values need no prefix rewriting on the frontend. 
     *
     * @param roomId  (required)
     * @param filename  (required)
     * @return The image file. (status code 200)
     *         or No such file (unknown room, unknown filename, or a path escaping the room&#39;s upload directory). (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/uploads/rooms/{roomId}/{filename}",
        produces = { "image/jpeg", "image/png", "image/webp", "application/json" }
    )
    
    default ResponseEntity<org.springframework.core.io.Resource> getRoomImage(
         @PathVariable("roomId") String roomId,
         @PathVariable("filename") String filename
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /public/rooms : List all rooms (public)
     * Unauthenticated. Powers the public room listing page. Nothing in &#x60;Room&#x60; is sensitive today, so it&#39;s reused as-is rather than duplicated into a narrower projection.
     *
     * @return All rooms. (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/public/rooms",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Room>> listPublicRooms(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" }, { \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"activeUnitCount\" : 0, \"images\" : [ \"images\", \"images\" ], \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"id\", \"capacity\" : 0, \"basePrice\" : \"basePrice\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}